\name{add_copy_object}
\alias{add_copy_object}
\title{Add a copy of an object to a model}
\usage{
  add_copy_object(compStr, obj, index)
}
\arguments{
  \item{compStr}{a string indicating the SBML component to
  be copied}

  \item{obj}{the model object to which the component will
  be added}

  \item{index}{a vector of integers indicating the
  positions of the objects to be copied}
}
\description{
  A function to add a copy of an SBML component object to a
  model
}

