\name{add_new_object}
\alias{add_new_object}
\title{Add a new object to a model}
\usage{
  add_new_object(crStr, obj)
}
\arguments{
  \item{crStr}{a string indicating the SBML component to be
  created}

  \item{obj}{the model object to which the component will
  be added}
}
\description{
  A function to add a new SBML component object to a model
}
\details{
  The new object will be initialised to typical defaults
}

