\docType{methods}
\name{barplot}
\alias{barplot}
\alias{barplot,_p_Model-method}
\alias{barplot,_p_Model,ANY-method}
\title{Barplot}
\arguments{
  \item{height}{a _p_Model object}
}
\description{
  A barplot for a _p_Model object
}
\details{
  Displays a barplot illustrating numbers of components
  within the model
}


