\name{convert_p_ListOf_to_R_ListOf}
\alias{convert_p_ListOf_to_R_ListOf}
\title{Change object class}
\usage{
  convert_p_ListOf_to_R_ListOf(obj)
}
\arguments{
  \item{obj}{the _p_ListOf object to be converted to an
  R_ListOf object}
}
\value{
  the libSBML _p_ListOf container as an R_ListOf object
}
\description{
  convert_p_ListOf_to_R_ListOf
}
\details{
  Converts a _p_ListOf object to an R_ListOf object
}

