\docType{methods}
\name{createFunctionDefinition}
\alias{createCompartment}
\alias{createCompartment,_p_Model-method}
\alias{createCompartment,_p_Model,ANY-method}
\alias{createCompartmentType}
\alias{createCompartmentType,_p_Model-method}
\alias{createCompartmentType,_p_Model,ANY-method}
\alias{createConstraint}
\alias{createConstraint,_p_Model-method}
\alias{createConstraint,_p_Model,ANY-method}
\alias{createEvent}
\alias{createEvent,_p_Model-method}
\alias{createEvent,_p_Model,ANY-method}
\alias{createFunctionDefinition}
\alias{createFunctionDefinition,_p_Model-method}
\alias{createFunctionDefinition,_p_Model,ANY-method}
\alias{createInitialAssignment}
\alias{createInitialAssignment,_p_Model-method}
\alias{createInitialAssignment,_p_Model,ANY-method}
\alias{createParameter}
\alias{createParameter,_p_Model-method}
\alias{createParameter,_p_Model,ANY-method}
\alias{createReaction}
\alias{createReaction,_p_Model-method}
\alias{createReaction,_p_Model,ANY-method}
\alias{createSpecies}
\alias{createSpecies,_p_Model-method}
\alias{createSpecies,_p_Model,ANY-method}
\alias{createSpeciesType}
\alias{createSpeciesType,_p_Model-method}
\alias{createSpeciesType,_p_Model,ANY-method}
\alias{createUnitDefinition}
\alias{createUnitDefinition,_p_Model-method}
\alias{createUnitDefinition,_p_Model,ANY-method}
\title{A generic for createFunctionDefinition}
\arguments{
  \item{obj}{a_p_Model object}

  \item{i}{a vector of integers indicating objects to be
  copied or created}
}
\description{
  A generic for createFunctionDefinition
}
\note{
  Where a new _p_* object is created, the SBML Level and
  Version of the _p_Model object, (and any SBML package
  namespaces), are used to initialise the attributes of the
  new _p_FunctionDefinition object to 'typical' defaults
}


