\docType{methods}
\name{edit_component}
\alias{[[<-,_p_Component,ANY-method}
\alias{[[<-,_p_Component,character,missing-method}
\alias{edit_component}
\title{Edit a _p_Component object attribute}
\arguments{
  \item{x}{a _p_Component object}

  \item{i}{a character string indicating the attribute to
  be changed}

  \item{value}{the new value of the attribute}
}
\description{
  Edit an attribute of a _p_Component object
}
\details{
  Changes the value of attribute
}

