\name{EventAssignment_attributes}
\alias{EventAssignment_attributes}
\title{EventAssignment component attributes}
\usage{
  EventAssignment_attributes()
}
\description{
  A listing of attributes of the libSBML 'EventAssignment'
  component
}
\note{
  The 'EventAssignment' component exists within the 'Event'
  component
}

