\name{get_attribute_from_component}
\alias{get_attribute_from_component}
\title{Component attributes}
\usage{
  get_attribute_from_component(compStr, obj, attrStr)
}
\arguments{
  \item{compStr}{a string containing the name of an SBML
  component (singular)}

  \item{obj}{the component object from which the attribute
  value will be taken}

  \item{attrStr}{a string containing the name of the
  attribute}
}
\value{
  the value of the attribute named in \code{attrStr} from
  the component named in \code{compStr}
}
\description{
  get_attribute_from_component
}
\details{
  Returns a component attribute
}

