\name{get_error_code}
\alias{get_error_code}
\title{libSBML error codes}
\usage{
  get_error_code(i)
}
\arguments{
  \item{i}{an integer value returned by a libSBML function}
}
\value{
  a string corresponding to the libSBML return value
  \code{i}
}
\description{
  A function to return an informative string corresponding
  to libSBML operation return value
}

