\name{get_math}
\alias{get_math}
\title{get_math}
\usage{
  get_math(compStr, obj)
}
\arguments{
  \item{compStr}{a string containing the name of an SBML
  component (singular)}

  \item{obj}{the component object containing the math
  element}
}
\value{
  a _p_ASTNode object
}
\description{
  get_math
}
\details{
  Get the math element of a component as a _p_ASTNode
  object
}

