\name{get_math_string_Constraint}
\alias{get_math_string_Constraint}
\title{Constraint math}
\usage{
  get_math_string_Constraint(con)
}
\arguments{
  \item{con}{a _p_Constraint object}
}
\value{
  a string describing the Constraint math
}
\description{
  get_math_string_Constraint
}
\details{
  Get an easily readable string describing math within the
  Constraint component
}

