\name{get_math_string_Delay}
\alias{get_math_string_Delay}
\title{Delay math}
\usage{
  get_math_string_Delay(del)
}
\arguments{
  \item{del}{a _p_Delay object}
}
\value{
  a string describing the Delay math
}
\description{
  get_math_string_Delay
}
\details{
  Get an easily readable string describing math within the
  Delay component
}
\note{
  Delay exists within the Event component
}

