\name{get_math_string_EventAssignment}
\alias{get_math_string_EventAssignment}
\title{EventAssignment math}
\usage{
  get_math_string_EventAssignment(ea)
}
\arguments{
  \item{ea}{a _p_EventAssignment object}
}
\value{
  a string describing the EventAssignment math
}
\description{
  get_math_string_EventAssignment
}
\details{
  Get an easily readable string describing math within the
  EventAssignment element
}
\note{
  EventAssignment exists within the Event component
}

