\name{get_math_string_FunctionDefinition}
\alias{get_math_string_FunctionDefinition}
\title{FunctionDefinition math}
\usage{
  get_math_string_FunctionDefinition(fd)
}
\arguments{
  \item{fd}{a _p_FunctionDefinition object}
}
\value{
  a string describing the FunctionDefinition math
}
\description{
  get_math_string_FunctionDefinition
}
\details{
  Get an easily readable string describing math within the
  FunctionDefinition component
}

