\name{get_math_string_InitialAssignment}
\alias{get_math_string_InitialAssignment}
\title{InitialAssignment math}
\usage{
  get_math_string_InitialAssignment(ia)
}
\arguments{
  \item{ia}{a _p_InitialAssignment object}
}
\value{
  a string describing the InitialAssignment math
}
\description{
  get_math_string_InitialAssignment
}
\details{
  Get an easily readable string describing math within the
  InitialAssignment component
}

