\name{get_math_string_KineticLaw}
\alias{get_math_string_KineticLaw}
\title{KineticLaw math}
\usage{
  get_math_string_KineticLaw(kl)
}
\arguments{
  \item{kl}{a _p_KineticLaw object}
}
\value{
  a string describing the KineticLaw math
}
\description{
  get_math_string_KineticLaw
}
\details{
  Get an easily readable string describing math within the
  KineticLaw component
}
\note{
  KineticLaw exists within the Reaction component
}

