\name{get_math_string_Priority}
\alias{get_math_string_Priority}
\title{Priority math}
\usage{
  get_math_string_Priority(p)
}
\arguments{
  \item{p}{a _p_Priority object}
}
\value{
  a string describing the Priority math
}
\description{
  get_math_string_Priority
}
\details{
  Get an easily readable string describing math within the
  Priority component
}
\note{
  Priority exists within the Event component
}

