\name{get_math_string_Rule}
\alias{get_math_string_Rule}
\title{Rule math}
\usage{
  get_math_string_Rule(r)
}
\arguments{
  \item{r}{a _p_Rule object}
}
\value{
  a string describing the Rule math
}
\description{
  get_math_string_Rule
}
\details{
  Get an easily readable string describing math within the
  Rule component
}

