\name{get_math_string_Trigger}
\alias{get_math_string_Trigger}
\title{Trigger math}
\usage{
  get_math_string_Trigger(tr)
}
\arguments{
  \item{tr}{a _p_Trigger object}
}
\value{
  a string describing the Trigger math
}
\description{
  get_math_string_Trigger
}
\details{
  Get an easily readable string describing math within the
  Trigger component
}
\note{
  Trigger exists within the Event component
}

