\name{get_R_ListOf_classes}
\alias{get_R_ListOf_classes}
\title{R_ListOf object classes}
\usage{
  get_R_ListOf_classes()
}
\description{
  A listing of container classes for libSBML component
  objects
}
\note{
  These are R 'list' objects containing libSBML component
  objects. Not all objects have container classes
}

