\name{libsbmlwrapper-package}
\alias{libsbmlwrapper-package}
\alias{libsbmlwrapper}
\docType{package}
\title{
libsbmlwrapper
}
\description{
Package provides R-like functions for working with SBML models
}
\details{
\tabular{ll}{
Package: \tab libsbmlwrapper\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-04\cr
License: \tab What license is it under?\cr
Depends: \tab methods\cr
}
Package enables SBML models to be loaded into R, and offers a range of functions for extracting and editing model attributes.
}
\author{
John Casement

Maintainer: John Casement
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{

}
\examples{

}
