\name{LocalParameter_attributes}
\alias{LocalParameter_attributes}
\title{LocalParameter component attributes}
\usage{
  LocalParameter_attributes()
}
\description{
  A listing of attributes of the libSBML 'LocalParameter'
  component
}
\note{
  The 'LocalParameter' component exists within the
  'KineticLaw' element of the 'Reaction' component
}

