\docType{class}
\name{R_ListOfFunctionDefinitions-class}
\alias{R_ListOfCompartments}
\alias{R_ListOfCompartments-class}
\alias{R_ListOfCompartmentTypes}
\alias{R_ListOfCompartmentTypes-class}
\alias{R_ListOfConstraints}
\alias{R_ListOfConstraints-class}
\alias{R_ListOfEventAssignments}
\alias{R_ListOfEventAssignments-class}
\alias{R_ListOfEvents}
\alias{R_ListOfEvents-class}
\alias{R_ListOfFunctionDefinitions-class}
\alias{R_ListOfInitialAssignments}
\alias{R_ListOfInitialAssignments-class}
\alias{R_ListOfLocalParameters}
\alias{R_ListOfLocalParameters-class}
\alias{R_ListOfModifierSpeciesReferences}
\alias{R_ListOfModifierSpeciesReferences-class}
\alias{R_ListOfParameters}
\alias{R_ListOfParameters-class}
\alias{R_ListOfReactions}
\alias{R_ListOfReactions-class}
\alias{R_ListOfRules}
\alias{R_ListOfRules-class}
\alias{R_ListOfSpecies}
\alias{R_ListOfSpecies-class}
\alias{R_ListOfSpeciesReferences}
\alias{R_ListOfSpeciesReferences-class}
\alias{R_ListOfSpeciesTypes}
\alias{R_ListOfSpeciesTypes-class}
\alias{R_ListOfUnitDefinitions}
\alias{R_ListOfUnitDefinitions-class}
\alias{R_ListOfUnits}
\alias{R_ListOfUnits-class}
\title{R_ListOfFunctionDefinitions class}
\description{
  R_ListOfFunctionDefinitions class
}
\note{
  an R 'list' containing libSBML _p_* objects
}
