\docType{methods}
\name{readSBML}
\alias{readSBML}
\title{Read an SBML file}
\usage{
  readSBML(filename)
}
\arguments{
  \item{filename}{a character string representing an SBML
  file}
}
\description{
  Read an SBML file
}
\details{
  Reads SBML file, sets defaults, and returns a
  _p_SBMLDocument object
}

