\docType{methods}
\name{readSBMLFromString}
\alias{readSBMLFromString}
\title{Read an SBML document from an xml string}
\usage{
  readSBMLFromString(xml)
}
\arguments{
  \item{xml}{string containing an SBML file}
}
\description{
  Read an SBML document from an xml string
}
\details{
  Reads SBML document, sets defaults, and returns a
  _p_SBMLDocument object
}

