\docType{methods}
\name{select_from_component}
\alias{[[,_p_Component,ANY-method}
\alias{[[,_p_Component,character,missing-method}
\alias{select_from_component}
\title{Selection of component attribute}
\arguments{
  \item{x}{a _p_Component object}

  \item{i}{a character string indicating the attribute
  being selected}
}
\value{
  a component attribute
}
\description{
  A select function for a _p_Component object
}
\details{
  Selects an attribute from an SBML component object
}

