\docType{methods}
\name{select_from_R_ListOf}
\alias{[,R_ListOf,ANY-method}
\alias{[,R_ListOf,numeric-method}
\alias{select_from_R_ListOf}
\title{Select a subset of a list of SBML component objects}
\arguments{
  \item{x}{an R_ListOf object}

  \item{i}{a vector indicating the components in the list
  to be selected}
}
\value{
  an R_ListOf object containing SBML component objects
}
\description{
  A select function for an R_ListOf object
}
\details{
  Selects a subset of an R_ListOf container object
}

