\name{select_from_Constraint}
\alias{select_from_Constraint}
\title{Select from _p_Constraint}
\usage{
  select_from_Constraint(x, i)
}
\arguments{
  \item{x}{a _p_Constraint object}

  \item{i}{a string which should be "Message"}
}
\value{
  the Constraint message string
}
\description{
  select_from_Constraint
}
\details{
  Handles requests for "Message" from a _p_Constraint
  object
}
\note{
  The constraint component has no attributes
}

