\docType{methods}
\name{show_ASTNode}
\alias{show,_p_ASTNode-method}
\alias{show,_p_ASTNode,ANY-method}
\alias{show_ASTNode}
\title{Show _p_ASTNode object}
\arguments{
  \item{object}{a _p_ASTNode object}
}
\description{
  A show function for an _p_ASTNode object
}
\details{
  Overrides generic 'show' method for a _p_ASTNode object
  and shows the math as a string

  Note that _p_ASTNode does not belong to the virtual
  superclass _p_Component
}

