\docType{methods}
\name{show_component}
\alias{show,_p_Component-method}
\alias{show,_p_Component,ANY-method}
\alias{show_component}
\title{Show _p_Component object}
\arguments{
  \item{object}{a _p_Component object}
}
\description{
  A show function for an _p_Component object
}
\details{
  Overrides generic 'show' method
}

