\docType{methods}
\name{show_document}
\alias{show}
\alias{show,_p_ASTNode-method}
\alias{show,_p_ASTNode,ANY-method}
\alias{show,_p_Component-method}
\alias{show,_p_Component,ANY-method}
\alias{show,_p_Model-method}
\alias{show,_p_Model,ANY-method}
\alias{show,_p_SBMLDocument-method}
\alias{show,_p_SBMLDocument,ANY-method}
\alias{show,R_ListOf-method}
\alias{show,R_ListOf,ANY-method}
\alias{show_document}
\title{Show document}
\arguments{
  \item{object}{a _p_SBMLDocument object}
}
\description{
  A show function for a _p_SBMLDocument object
}
\details{
  Overrides generic 'show' method
}

