\docType{methods}
\name{show_RListOf}
\alias{show,R_ListOf-method}
\alias{show,R_ListOf,ANY-method}
\alias{show_RListOf}
\title{Show R_ListOf object}
\arguments{
  \item{object}{an R_ListOf object}
}
\description{
  A show function for an R_ListOf object
}
\details{
  Overrides generic 'show' method
}

