\name{SimpleSpeciesReference_attributes}
\alias{SimpleSpeciesReference_attributes}
\title{SimpleSpeciesReference component attributes}
\usage{
  SimpleSpeciesReference_attributes()
}
\description{
  A listing of attributes of the libSBML
  'SimpleSpeciesReference' component
}
\note{
  The 'SimpleSpeciesReference' component exists within the
  'Reaction' component
}

