\name{SpeciesReference_attributes}
\alias{SpeciesReference_attributes}
\title{SpeciesReference component attributes}
\usage{
  SpeciesReference_attributes()
}
\description{
  A listing of attributes of the libSBML 'SpeciesReference'
  component
}
\note{
  The 'SpeciesReference' component exists within the
  'Reaction' component
}

