\docType{methods}
\name{summary_component}
\alias{summary,_p_Component-method}
\alias{summary,_p_Component,ANY-method}
\alias{summary_component}
\title{Component summary}
\arguments{
  \item{object}{an SBML component}
}
\description{
  A summary function for a _p_Component object
}
\details{
  Displays values of attributes of an SBML component object
}

