\docType{methods}
\name{summary_model}
\alias{summary,_p_Model-method}
\alias{summary,_p_Model,ANY-method}
\alias{summary_model}
\title{Model summary}
\arguments{
  \item{object}{an SBML model}
}
\description{
  A summary function for a _p_Model object
}
\details{
  Returns a list containing attributes of the model
}

