\docType{methods}
\name{summary_document}
\alias{summary,_p_SBMLDocument-method}
\alias{summary,_p_SBMLDocument,ANY-method}
\alias{summary_document}
\title{Document summary}
\arguments{
  \item{object}{an SBML document}
}
\description{
  A summary function for a _p_SBMLDocument object
}
\details{
  Returns a list containing high level information about
  the model
}

