\docType{methods}
\name{show_document}
\alias{show,_p_SBMLDocument-method}
\alias{show,_p_SBMLDocument,ANY-method}
\alias{show_document}
\title{Show document}
\arguments{
  \item{object}{a _p_SBMLDocument object}
}
\description{
  A show function for a _p_SBMLDocument object
}
\details{
  Overrides generic 'show' method
}

