\docType{methods}
\name{summary_document}
\alias{summary}
\alias{summary,_p_Component-method}
\alias{summary,_p_Component,ANY-method}
\alias{summary,_p_Model-method}
\alias{summary,_p_Model,ANY-method}
\alias{summary,_p_SBMLDocument-method}
\alias{summary,_p_SBMLDocument,ANY-method}
\alias{summary,R_ListOf-method}
\alias{summary,R_ListOf,ANY-method}
\alias{summary_document}
\title{Summary methods}
\description{
  S4 summary methods for _p_SBMLDocument, _p_model,
  _p_Component and R_List objects.
}

